MOM-561 Upgrade winston version
The previous code was breaking with the latest version of winston.
I suspect that the way the latest versions of winston initializes the logger
has changed and therefore no longer compitable with older versions.

This commit includes updating all winston references to use the newly
created logger wrapper module for winston.

Related Tickets:
 - MOM-519